Moved login to LoginContainer
